big update
another big update to Stryke, this time we are accounting for the fact that a fish may or may not be sampled that day - not all fish present at a facility are captured during every sampling event - that means we have information on an absence and we can flip a coin weighted with probability of occurence.  New example spreadsheet as well as some summary queries.  Also new updated EPRI database to conform to our presence/absence.  Also made the simulation and I/O wicked fast and made the EPRI query script output easier to read.